Fix building of golua against a shared Lua-5.1.4
Currently cgo does not seem to be able to link against the liblua.a that
is provided by the standard Lua installation. You will need to build a
liblua.so and install that in your system in order for this binding to
work. Hopefully I can remove this limitation and create a standalone
static golua, but I'm unsure how to do that at this time.