- take out local lua build (breaks my build) - use dynamic library instead. - use system /usr/include/lua5.1 (probably should use pkg-config or similar) - remove comment in readme
did not want -- the version on my system is the version I want to bind against,
and I see no compelling reason to keep a whole tree inline.

To include alternate headers/dylib, modify the makefile defaults.
For that matter, you'll probably need to tweak them for anything that's
not my exact version of debian sid, but this is my branch :)